requirements.txt
streamlit
pandas
plotly
openpyxl
psycopg2-binary
python-dotenv
